refactor(livewire): reorganize components by feature and update Blade aliases
