+++ reservations/app/Livewire/Auth/RegisterForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Livewire;
+namespace App\Livewire\Auth;
 
 use App\Livewire\Forms\RegisterFormObject;
 use App\Models\User;
