+++ reservations/app/Livewire/Home/FeaturedList.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Livewire;
+namespace App\Livewire\Home;
 
 use App\Models\Restaurant;
 use Livewire\Component;
@@ -24,7 +24,7 @@ class FeaturedList extends Component
 
     public function render()
     {
-        return view('livewire.featured-list', [
+        return view('livewire.home.featured-list', [
             "restaurants" => Restaurant::orderBy('rating', 'desc')->take(3)->get(),
         ]);
     }
