+++ reservations/app/Livewire/Home/SearchResult.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Livewire;
+namespace App\Livewire\Home;
 
 use App\Models\Restaurant;
 use Livewire\Component;
@@ -13,7 +13,8 @@ class SearchResult extends Component
     public $search = "";
     protected $listeners = ['restaurantSearchUpdated' => 'setSearch'];
 
-    public function setSearch($term) {
+    public function setSearch($term)
+    {
         $this->search = $term;
         $this->resetPage();
     }
@@ -27,7 +28,7 @@ class SearchResult extends Component
             ->orderBy('rating', 'desc')
             ->paginate(10);
 
-        return view('livewire.search-result', [
+        return view('livewire.home.search-result', [
             'restaurants' => $restaurants,
         ]);
     }
