+++ reservations/resources/views/livewire/home/featured-list.blade.php
