+++ reservations/resources/views/pages/auth/register-page.blade.php
@@ -1,7 +1,7 @@
 <x-layouts.app>
     <div class="inset-0 flex items-center justify-center mt-10">
         <div class="w-full max-w-md rounded-2xl bg-neutral-900 p-8 shadow-2xl">
-            <livewire:register-form />
+            <livewire:auth.register-form />
 
             <div class="mt-6 text-center text-sm text-neutral-400">
                 Already have an account?
